Time: 52 ms (22.44%), Space: 55.8 MB (86.54%) - LeetHub

+++ 2365-task-scheduler-ii/2365-task-scheduler-ii.java
@@ -5,7 +5,7 @@ class Solution {
         
         for(int ptr=0; ptr<tasks.length; ++ptr){
             if(map.containsKey(tasks[ptr]) && map.get(tasks[ptr])>time){
-                time = (long)map.get(tasks[ptr]);
+                time = map.get(tasks[ptr]);
                 ptr--;
             }
             else{
